Say 'template'.

+++ quantities/parser_body.hpp
@@ -220,7 +220,7 @@ inline Unit ParseQuotientUnit(std::string const& s) {
     // input like 1.23 / s.
     int const first_nonblank = s.find_first_not_of(' ', last_slash + 1);
     CHECK_NE(std::string::npos, first_nonblank);
-    int const last_nonblank = last_slash == 0
+    size_t const last_nonblank = last_slash == 0
                                   ? std::string::npos
                                   : s.find_last_not_of(' ', last_slash - 1);
     auto const left = last_nonblank == std::string::npos
